Add bid functionality (#11)
* Update .gitignore

* Update .gitignore

* Update app.py

* Update init.sql

* Update init.sql

* Update init.sql

* Update init.sql

* Update init.sql

* Update init.sql

* Update init.sql

* Update init.sql

* Correct indentation issues

* Update app.py

* Update init.sql

* Update init.sql

* Update app.py

* Update init.sql

* Remove license number from driver

* Update init.sql

* Working validation code

* kaboom

* Working code for adding/updating bids

* Added working code for deleting bid

* Update home.html

* Fix bug

* Remove blank lines

+++ init.sql
@@ -86,7 +86,7 @@ CREATE TABLE Bids (
     passenger_ID     varchar(50) REFERENCES Passenger ON DELETE CASCADE,
     driver_ID        varchar(50) REFERENCES Driver    ON DELETE CASCADE,
     time_posted      TIMESTAMP,
-    price            INTEGER,
+    price            NUMERIC,
     status           varchar(20),
     no_passengers    INTEGER,
     PRIMARY KEY (passenger_ID, time_posted, driver_ID),
